:recycle: [REFAC] 홈 화면 리팩터링
